fix comment component
